+++ client-src/components.js
@@ -41,6 +41,7 @@ registerIconLibrary('material', {
 
 // chromedash components
 import './elements/icons';
+import './elements/chromedash-admin-blink-component-listing';
 import './elements/chromedash-admin-blink-page';
 import './elements/chromedash-all-features-page';
 import './elements/chromedash-activity-log';
